Chrome: Move local state parsing into ChromeBrowserLocalData
